set up for resume
modify the data through resumeJSON.py.
embed it in the html with RESUMEJSON.js.
delete hard coded data from resumetemplate.html.
modify the js with iterating over the data so that it isnt restricted to a certain number of elements.
end up with a dynamic resume environment, in which you can simply change everything about the resume file within python.